Added key list controller

+++ src/ntk/utils/convert.py
@@ -5,10 +5,10 @@ class Convert:
     """Handles converting units of measurement."""
 
     @staticmethod
-    def to_cm(inches: int) -> float:
+    def to_cm(inches: float) -> float:
         """Convert inches to cm.
 
-        :param inches: number of inches (int)
+        :param inches: number of inches (float)
         :return: cm conversion (float)
         """
         return round(float(inches) * 2.54, 2)
@@ -23,10 +23,10 @@ class Convert:
         return round(float(lbs) / 2.2, 2)
 
     @staticmethod
-    def to_meters(inches: int) -> float:
+    def to_meters(inches: float) -> float:
         """Convert inches to meters.
 
-        :param inches: number of inches (int)
+        :param inches: number of inches (float)
         :return: meters (float)
         """
         return round(float(inches) * 0.0254, 2)
